moved info setting functions into folder within fill_users
